requirements.txt
requests
pandas
tqdm
pyyaml
